copilot-metrics-functions/requirements.txt
azure-functions
pyodbc